fix(models): electrostatics gradients and DD-pure specs
Ewald's reciprocal term was only ever differentiated with respect to
positions and charges, so a non-hybrid Ewald returned a wrong `dE/dq`, and
strain-autograd through the detached Green's function gave a wrong stress. Work
needing a cell derivative now routes to the staged reciprocal.

Cached k-vectors were rebuilt from the strained cell, so a second stress
evaluation reused the first call's autograd graph.

Ewald and PME now default to `hybrid_forces=False`; the analytic direct-output
path does not produce consistent gradients and is rejected under domain
decomposition. Their halo and graph-parallel specs declare stress.

Requires nvalchemi-toolkit-ops 0.4.1.

+++ test/distributed/model/test_ewald_multigpu.py
@@ -213,7 +213,7 @@ def _ewald_equivalence_worker(rank: int, world_size: int) -> None:
     # energy holds to ~1e-4 absolute; per-atom forces to ~1e-3 relative
     # (same tolerance the cueq/MACE multi-GPU test uses).
     torch.testing.assert_close(
-        e_local.view(1),
+        e_local.view(1).to(e_ref.dtype),
         e_ref,
         rtol=1e-4,
         atol=1e-4,
