+++ test/distributed/model/test_pme_multigpu.py
@@ -209,7 +209,7 @@ def _pme_equivalence_worker(rank: int, world_size: int) -> None:
     # fp32 + FFT-based PME: tolerances slightly looser than Ewald's
     # direct k-sum because of accumulated rounding in the mesh pipeline.
     torch.testing.assert_close(
-        e_local.view(1),
+        e_local.view(1).to(e_ref.dtype),
         e_ref,
         rtol=5e-4,
         atol=5e-4,
